@@ -28,7 +28,7 @@ class RegisterController extends Controller
      *
      * @var string
      */
-    protected $redirectTo ='/';
+    protected string $redirectTo ='/';
 
     /**
      * Create a new controller instance.
