@@ -5,7 +5,7 @@ namespace App\Constants;
 class AppConstants
 {
     public const QUEUE_REDIS_READ_WRITE_TEST = 'read-write-test';
-    public const CHANNEL_REDIS_READ_WRITE_TEST = 'REDIS_READ_WRITE_TEST';
+    public const CHANNEL_REDIS_READ_WRITE_TEST = 'redis-read-write-test';
     public const QUEUE_LONG_RUNNING = 'long-running-queue';
 
 
