@@ -37,7 +37,7 @@ class Handler extends ExceptionHandler
      *
      * @throws Exception|Throwable
      */
-    public function report(Throwable $exception)
+    public function report(Throwable $exception): void
     {
         parent::report($exception);
     }
